% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{add_workdays}
\alias{add_workdays}
\title{计算时间段内包含的工作日数量}
\usage{
add_workdays(
  d,
  from,
  to,
  newName = "工作日数",
  path = NULL,
  ignoreFrom = FALSE,
  ignoreTo = FALSE,
  withSeq = FALSE
)
}
\arguments{
\item{d}{要补充字段的数据集}

\item{from}{开始日期字段名}

\item{to}{截止日期字段名}

\item{newName}{工作日字段命名}

\item{path}{指定节假日设定文件}

\item{ignoreFrom}{计算工作日时忽略开始列}

\item{ignoreTo}{计算工作日时忽略结束列}

\item{withSeq}{返回工作日的标记序列，以便后续计算}
}
\description{
计算时间段内包含的工作日数量
}
