% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays-package.R
\docType{package}
\name{workdays-package}
\alias{workdays}
\alias{workdays-package}
\title{workdays: count workdays before two date columns}
\description{
count workdays before two date columns.
}
\author{
\strong{Maintainer}: hongwei xue \email{43801@qq.com}

}
\keyword{internal}
