% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{workdays_count}
\alias{workdays_count}
\title{计算时间段内包含的工作日数量}
\usage{
workdays_count(
  d,
  fromColumn,
  toColumn,
  idColumn = NULL,
  seqIgnore = NULL,
  newName = "工作日数",
  fromIgnore = FALSE,
  toIgnore = FALSE,
  path = NULL,
  withRaw = FALSE
)
}
\arguments{
\item{d}{要处理的数据集}

\item{fromColumn}{开始日期字段名}

\item{toColumn}{截止日期字段名}

\item{idColumn}{合并工作日依据的ID字段名称}

\item{newName}{工作日字段命名}

\item{fromIgnore}{计算工作日时忽略开始列}

\item{toIgnore}{计算工作日时忽略结束列}

\item{path}{指定节假日设定文件}
}
\description{
支持按照ID字段合并
}
