% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{workdays_seq}
\alias{workdays_seq}
\title{计算时间段内包含的工作日序列}
\usage{
workdays_seq(
  d,
  fromColumn,
  toColumn,
  idColumn,
  fromIgnore = TRUE,
  toIgnore = TRUE
)
}
\arguments{
\item{d}{要处理的数据集}

\item{fromColumn}{开始日期字段名}

\item{toColumn}{截止日期字段名}

\item{idColumn}{合并工作日依据的ID字段名称}

\item{fromIgnore}{计算工作日时忽略开始列}

\item{toIgnore}{计算工作日时忽略结束列}
}
\description{
如果需要按照ID排除工作日数量，则要求调用该函数。
}
