% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{workdays_zh_CN}
\alias{workdays_zh_CN}
\title{工作日清单}
\usage{
workdays_zh_CN(fromDay = "2023-01-01", toDay = "2023-12-31", path = NULL)
}
\arguments{
\item{fromDay}{开始日期}

\item{toDay}{截止日期}
}
\description{
默认为周一至周五为工作日，再根据中国国务院发布的工作日调整政策修正
}
\concept{workday-funcs}
