% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{add_workdays}
\alias{add_workdays}
\title{计算时间段内包含的工作日数量}
\usage{
add_workdays(d, from, to, newName = "工作日数", id = NULL)
}
\arguments{
\item{d}{要补充字段的数据集}

\item{newName}{工作日字段命名}

\item{fromColumn}{开始日期字段名}

\item{toColumn}{截止日期字段名}
}
\description{
计算时间段内包含的工作日数量
}
