% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdays.R
\name{add_workdays_by_id}
\alias{add_workdays_by_id}
\title{计算时间段内包含的工作日数量，按照ID合并}
\usage{
add_workdays_by_id(d, from, to, idName, newName = "工作日数")
}
\arguments{
\item{d}{要补充字段的数据集}

\item{idName}{合并工作日依据的ID字段名称}

\item{newName}{工作日字段命名}

\item{fromColumn}{开始日期字段名}

\item{toColumn}{截止日期字段名}
}
\description{
计算时间段内包含的工作日数量，按照ID合并
}
